added 10 addins to test tommorow

+++ Processors/ExportDxfsProcessor.cs
@@ -6,14 +6,15 @@ namespace SolidEdgeAdd_In.Processors
     {
         private readonly SeAssembly _assembly;
         private readonly SeApp _application;
-        private readonly Logger _logger;
 
-        private string _assemblyFullName;
+        private readonly Logger _logger;
+        private readonly Dictionary<string, FileData> _occurrencesData;
+
+        private string _assemblyFilePath;
         private string _assemblyDirectory;
 
         private int _multiplier;
-        private string _subDirectory;
-        private Dictionary<string, FileData> _occurrencesData;
+        private string _targetDirectory;
 
         public ExportDxfsProcessor(SeAssembly assembly)
         {
@@ -25,15 +26,12 @@ namespace SolidEdgeAdd_In.Processors
 
         public bool Initialize()
         {
-            _assemblyFullName = _assembly.FullName;
-            _assemblyDirectory = Path.GetDirectoryName(_assemblyFullName);
+            _assemblyFilePath = _assembly.FullName;
+            _assemblyDirectory = Path.GetDirectoryName(_assemblyFilePath);
 
-            if (!TryGetMultiplier())
-            {
-                return false;
-            }
+            if (!TryGetMultiplier()) { return false; }
 
-            SetSubDirectory();
+            SetTargetDirectory();
             LoadOccurrencesData();
 
             if (_occurrencesData.Count == 0)
@@ -58,6 +56,7 @@ namespace SolidEdgeAdd_In.Processors
             ExcelWorkbook workbook = null;
             ExcelSheets xlSheets = null;
             ExcelWorksheet worksheet = null;
+
             ExcelRange headerRange = null;
             ExcelRange usedRange = null;
             ExcelRange columns = null;
@@ -70,18 +69,9 @@ namespace SolidEdgeAdd_In.Processors
             dynamic headerBorders = null;
             dynamic usedBorders = null;
 
-            string mainDirectory = _assemblyDirectory;
-
             try
             {
-                excelApp = new ExcelApp
-                {
-                    Visible = false,
-                    DisplayAlerts = false,
-                    ScreenUpdating = false,
-                    EnableEvents = false
-                };
-
+                excelApp = new ExcelApp { Visible = false, DisplayAlerts = false, ScreenUpdating = false, EnableEvents = false };
                 workbooks = excelApp.Workbooks;
                 workbook = workbooks.Add();
                 excelApp.Calculation = Microsoft.Office.Interop.Excel.XlCalculation.xlCalculationManual;
@@ -93,17 +83,17 @@ namespace SolidEdgeAdd_In.Processors
                 {
                     {
                         Constants.ExcelHeaders.Lp,
-                        Constants.ExcelHeaders.PartNumber,
-                        Constants.ExcelHeaders.Quantity,
-                        Constants.ExcelHeaders.Name,
+                        Constants.ExcelHeaders.FileName,
+                        Constants.ExcelHeaders.Count,
+                        Constants.ExcelHeaders.Title,
                         Constants.ExcelHeaders.Thickness,
-                        Constants.ExcelHeaders.Width,
-                        Constants.ExcelHeaders.Length,
+                        Constants.ExcelHeaders.SizeX,
+                        Constants.ExcelHeaders.SizeY,
                         Constants.ExcelHeaders.Color,
                         Constants.ExcelHeaders.Finish,
                         Constants.ExcelHeaders.Flanges,
                         Constants.ExcelHeaders.Material,
-                        Constants.ExcelHeaders.Date
+                        Constants.ExcelHeaders.DxfDate
                     }
                 };
 
@@ -114,11 +104,7 @@ namespace SolidEdgeAdd_In.Processors
                     headerRange = worksheet.Range[startHeaderCell, endHeaderCell];
                     headerRange.Value = headerData;
                 }
-                finally
-                {
-                    CoreUtils.ReleaseCom(ref startHeaderCell);
-                    CoreUtils.ReleaseCom(ref endHeaderCell);
-                }
+                finally { Helpers.ReleaseCom(ref startHeaderCell); Helpers.ReleaseCom(ref endHeaderCell); }
 
                 headerFont = headerRange.Font;
                 headerFont.Bold = true;
@@ -132,55 +118,38 @@ namespace SolidEdgeAdd_In.Processors
 
                 var sortedData = _occurrencesData.OrderBy(item => item.Value.Material)
                      .ThenBy(item => double.TryParse(item.Value.Thickness?.Replace('_', ','), out double t) ? t : 0)
-                     .ThenBy(item => item.Value.Name)
-                     .ToList();
+                     .ThenBy(item => item.Value.FileName).ToList();
 
-                int i = 1;
+                int lp = 1;
                 foreach (var item in sortedData)
                 {
-                    string name = item.Value.Name;
+                    string fileName = item.Value.FileName;
                     int count = item.Value.OccurrenceCount * _multiplier;
                     string material = item.Value.Material;
                     string thickness = item.Value.Thickness;
 
-                    string mainDxFilefName = $"{thickness}mm_{material}_{name}.dxf";
-                    string subDxfFileName = $"{thickness}mm_{count}pcs_{material}_{name}.dxf";
+                    string mainDxfFileName = $"{thickness}mm_{material}_{fileName}.dxf";
+                    string subDxfFileName = $"{thickness}mm_{count}szt_{material}_{fileName}.dxf";
 
-                    string mainDxfPath = Path.Combine(mainDirectory, mainDxFilefName);
-                    string subDxfPath = Path.Combine(_subDirectory, subDxfFileName);
+                    string mainDxfPath = Path.Combine(_assemblyDirectory, mainDxfFileName);
+                    string subDxfPath = Path.Combine(_targetDirectory, subDxfFileName);
 
                     object finalSizeX = "-";
                     string sizeX = item.Value.SizeX;
-
                     if (!string.IsNullOrEmpty(sizeX))
                     {
                         sizeX = sizeX.ToLower().Replace("mm", "").Replace(" ", "").Replace(".", ",").Trim();
-
-                        if (double.TryParse(sizeX, out double parsedX))
-                        {
-                            finalSizeX = parsedX;
-                        }
-                        else
-                        {
-                            finalSizeX = sizeX;
-                        }
+                        if (double.TryParse(sizeX, out double parsedX)) { finalSizeX = parsedX; }
+                        else { finalSizeX = sizeX; }
                     }
 
                     object finalSizeY = "-";
                     string sizeY = item.Value.SizeY;
-
                     if (!string.IsNullOrEmpty(sizeY))
                     {
                         sizeY = sizeY.ToLower().Replace("mm", "").Replace(" ", "").Replace(".", ",").Trim();
-
-                        if (double.TryParse(sizeY, out double parsedY))
-                        {
-                            finalSizeY = parsedY;
-                        }
-                        else
-                        {
-                            finalSizeY = sizeY;
-                        }
+                        if (double.TryParse(sizeY, out double parsedY)) { finalSizeY = parsedY; }
+                        else { finalSizeY = sizeY; }
                     }
 
                     string color = string.IsNullOrEmpty(item.Value.Color) ? "-" : item.Value.Color;
@@ -189,65 +158,49 @@ namespace SolidEdgeAdd_In.Processors
 
                     string dxfDate = item.Value.DxfDate;
                     bool needGenerationDxf = string.IsNullOrEmpty(dxfDate);
-
                     bool isOpen = false;
                     string flangesValue = "-";
 
                     try
                     {
-                        document = CoreUtils.GetOpenDocument(_application, item.Key);
+                        document = Helpers.GetOpenDocument(_application, item.Key);
                         isOpen = true;
 
-                        if (document is SePart part)
-                        {
-                            models = part.Models;
-                            flatPatterns = part.FlatPatternModels;
-                        }
-                        else if (document is SeSheetMetal sheetMetal)
-                        {
-                            models = sheetMetal.Models;
-                            flatPatterns = sheetMetal.FlatPatternModels;
-                        }
+                        if (document is SePart part) { models = part.Models; flatPatterns = part.FlatPatternModels; }
+                        else if (document is SeSheetMetal sheetMetal) { models = sheetMetal.Models; flatPatterns = sheetMetal.FlatPatternModels; }
 
                         if (flatPatterns == null || models == null || flatPatterns.Count == 0 || models.Count == 0)
                         {
-                            _logger.LogSkip(name, "Missing flat pattern.");
+                            _logger.LogSkip(fileName, "Missing flat pattern.");
                             continue;
                         }
 
                         model = models.Item(1);
                         flanges = model.Flanges;
 
-                        if (flanges.Count != 0)
-                        {
-                            flangesValue = flanges.Count.ToString();
-                        }
+                        if (flanges.Count != 0) { flangesValue = flanges.Count.ToString(); }
 
                         if (needGenerationDxf)
                         {
-                            if (File.Exists(subDxfPath))
-                            {
-                                try { File.Delete(subDxfPath); } catch { }
-                            }
+                            if (File.Exists(subDxfPath)) { try { File.Delete(subDxfPath); } catch { } }
 
-                            using var properties = new PropertyProvider(document);
+                            using var properties = new PropertyUtils(document);
                             properties.UpdateDxfDate();
                             dxfDate = properties.DxfDate;
 
                             models.SaveAsFlatDXFEx(subDxfPath, null, null, null, true);
-                            _logger.LogSuccess($"{name} - Created new flat pattern Dxf.");
+
+                            document.Save();
+                            _logger.LogSuccess($"{fileName} - Created new flat pattern Dxf.");
                         }
-                        else
-                        {
-                            _logger.LogSuccess($"{name} - File has Dxf property.");
-                        }
+                        else { _logger.LogSuccess($"{fileName} - File has Dxf property."); }
 
                         if (File.Exists(subDxfPath))
                         {
                             File.Copy(subDxfPath, mainDxfPath, true);
 
-                            excelData[dataRowIndex, 0] = i++;
-                            excelData[dataRowIndex, 1] = name;
+                            excelData[dataRowIndex, 0] = lp++;
+                            excelData[dataRowIndex, 1] = fileName;
                             excelData[dataRowIndex, 2] = count;
                             excelData[dataRowIndex, 3] = title;
                             excelData[dataRowIndex, 4] = $"{thickness} mm";
@@ -261,34 +214,19 @@ namespace SolidEdgeAdd_In.Processors
 
                             dataRowIndex++;
                         }
-                        else
-                        {
-                            _logger.LogError(name, "File has Dxf property, but physical file not found.");
-                        }
+                        else { _logger.LogError(fileName, "File has Dxf property, but physical file not found."); }
                     }
-                    catch (Exception ex)
-                    {
-                        _logger.LogError(name, $"Error: {ex.Message}");
-                        continue;
-                    }
+                    catch (Exception ex) { _logger.LogError(fileName, $"Error: {ex.Message}"); continue; }
                     finally
                     {
-                        CoreUtils.ReleaseCom(ref flanges);
-                        CoreUtils.ReleaseCom(ref model);
-                        CoreUtils.ReleaseCom(ref flatPatterns);
-                        CoreUtils.ReleaseCom(ref models);
+                        Helpers.ReleaseCom(ref flanges); Helpers.ReleaseCom(ref model);
+                        Helpers.ReleaseCom(ref flatPatterns); Helpers.ReleaseCom(ref models);
 
                         if (isOpen && document != null)
                         {
-                            try
-                            {
-                                document.Save();
-                                document.Close(true);
-                            }
-                            catch { }
+                            try { document.Close(true); } catch { }
                         }
-
-                        CoreUtils.ReleaseCom(ref document);
+                        Helpers.ReleaseCom(ref document);
                     }
                 }
 
@@ -297,7 +235,6 @@ namespace SolidEdgeAdd_In.Processors
                     ExcelRange startCell = null;
                     ExcelRange endCell = null;
                     ExcelRange writeRange = null;
-
                     try
                     {
                         startCell = (ExcelRange)cells[2, 1];
@@ -305,12 +242,7 @@ namespace SolidEdgeAdd_In.Processors
                         writeRange = worksheet.Range[startCell, endCell];
                         writeRange.Value = excelData;
                     }
-                    finally
-                    {
-                        CoreUtils.ReleaseCom(ref writeRange);
-                        CoreUtils.ReleaseCom(ref endCell);
-                        CoreUtils.ReleaseCom(ref startCell);
-                    }
+                    finally { Helpers.ReleaseCom(ref writeRange); Helpers.ReleaseCom(ref endCell); Helpers.ReleaseCom(ref startCell); }
                 }
 
                 usedRange = worksheet.UsedRange;
@@ -321,63 +253,37 @@ namespace SolidEdgeAdd_In.Processors
                 usedBorders = usedRange.Borders;
                 usedBorders.LineStyle = Microsoft.Office.Interop.Excel.XlLineStyle.xlContinuous;
 
-                string excelPath = Path.Combine(_subDirectory, "DXF_Summary.xlsx");
+                string excelFilePath = Path.Combine(_targetDirectory, "DXF_Summary.xlsx");
 
-                if (File.Exists(excelPath))
-                {
-                    File.Delete(excelPath);
-                }
-
-                workbook.SaveAs(excelPath);
+                if (File.Exists(excelFilePath)) { File.Delete(excelFilePath); }
+                workbook.SaveAs(excelFilePath);
             }
             finally
             {
-                CoreUtils.ReleaseCom(ref usedBorders);
-                CoreUtils.ReleaseCom(ref headerBorders);
-                CoreUtils.ReleaseCom(ref headerInterior);
-                CoreUtils.ReleaseCom(ref headerFont);
-                CoreUtils.ReleaseCom(ref cells);
-                CoreUtils.ReleaseCom(ref columns);
-                CoreUtils.ReleaseCom(ref usedRange);
-                CoreUtils.ReleaseCom(ref headerRange);
-                CoreUtils.ReleaseCom(ref worksheet);
-                CoreUtils.ReleaseCom(ref xlSheets);
+                Helpers.ReleaseCom(ref usedBorders); Helpers.ReleaseCom(ref headerBorders);
+                Helpers.ReleaseCom(ref headerInterior); Helpers.ReleaseCom(ref headerFont);
+                Helpers.ReleaseCom(ref cells); Helpers.ReleaseCom(ref columns);
+                Helpers.ReleaseCom(ref usedRange); Helpers.ReleaseCom(ref headerRange);
+                Helpers.ReleaseCom(ref worksheet); Helpers.ReleaseCom(ref xlSheets);
 
-                if (workbook != null)
-                {
-                    try { workbook.Close(false); } catch { }
-                }
+                if (workbook != null) { try { workbook.Close(false); } catch { } }
+                Helpers.ReleaseCom(ref workbook); Helpers.ReleaseCom(ref workbooks);
 
-                CoreUtils.ReleaseCom(ref workbook);
-                CoreUtils.ReleaseCom(ref workbooks);
-
-                if (excelApp != null)
-                {
-                    try { excelApp.Quit(); } catch { }
-                }
-
-                CoreUtils.ReleaseCom(ref excelApp);
+                if (excelApp != null) { try { excelApp.Quit(); } catch { } }
+                Helpers.ReleaseCom(ref excelApp);
             }
         }
 
         private bool TryGetMultiplier()
         {
             SeDocument document = (SeDocument)_assembly;
-
-            using var properties = new PropertyProvider(document);
+            using var properties = new PropertyUtils(document);
             int count = properties.Count;
 
             if (count == 0)
             {
-                var (isConfirmed, multiplier) = DialogService.GetMultiplier();
-
-                if (isConfirmed)
-                {
-                    properties.Count = multiplier;
-                    _multiplier = multiplier;
-                    return true;
-                }
-
+                var (isConfirmed, multiplier) = DialogUtils.GetMultiplier();
+                if (isConfirmed) { properties.Count = multiplier; _multiplier = multiplier; return true; }
                 return false;
             }
 
@@ -385,40 +291,28 @@ namespace SolidEdgeAdd_In.Processors
             return true;
         }
 
-        private void SetSubDirectory()
+        private void SetTargetDirectory()
         {
-            string fileName = Path.GetFileNameWithoutExtension(_assemblyFullName);
-            string number = fileName.Length >= 4 ? fileName.Substring(0, 4) : fileName;
+            string assemblyFileName = Path.GetFileNameWithoutExtension(_assemblyFilePath);
+            string number = assemblyFileName.Length >= 4 ? assemblyFileName.Substring(0, 4) : assemblyFileName;
             string date = DateTime.Now.ToString("yyyy-MM-dd");
 
             string packagesDirectory = Path.Combine(_assemblyDirectory, Constants.Folders.Packages);
+            if (!Directory.Exists(packagesDirectory)) { Directory.CreateDirectory(packagesDirectory); }
 
-            if (!Directory.Exists(packagesDirectory))
-            {
-                Directory.CreateDirectory(packagesDirectory);
-            }
-
-            _subDirectory = Path.Combine(packagesDirectory, $"{number}_{date}");
-
-            if (!Directory.Exists(_subDirectory))
-            {
-                Directory.CreateDirectory(_subDirectory);
-            }
+            _targetDirectory = Path.Combine(packagesDirectory, $"{number}_{date}");
+            if (!Directory.Exists(_targetDirectory)) { Directory.CreateDirectory(_targetDirectory); }
         }
 
         private void LoadOccurrencesData()
         {
             SeOccurrences occurrences = null;
-
             try
             {
                 occurrences = _assembly.Occurrences;
-                AssemblyTreeWalker.BuildDataForExportDxfs(occurrences, _occurrencesData, _logger);
+                DataUtils.BuildDataForExportDxfs(occurrences, _occurrencesData, _logger);
             }
-            finally
-            {
-                CoreUtils.ReleaseCom(ref occurrences);
-            }
+            finally { Helpers.ReleaseCom(ref occurrences); }
         }
     }
 }
